backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pydantic
python-dotenv
passlib[bcrypt]
python-jose
psycopg2-binary
pandas
scikit-learn
joblib
numpy